Settings/Wifi: Add support for Ad-hoc (IBSS) networks
If the WLAN driver can support IBSS mode, show Ad-Hoc (IBSS) networks in the
list of available Wi-Fi networks and allow to connect to or create new Ad-Hoc
networks.

This is merged and ported from CM-10.2, where it was split up into 3 commits.

Change-Id: Ib58d05d05ce00a77d5b88ee57007d38798d7b43d

Conflicts:
	src/com/android/settings/wifi/AccessPoint.java
Why the heck not get another feature?